Parse corequisites (#53)
* add coreqs to crawler
