+++ src/constants.ts
@@ -1,0 +1,120 @@
+/**
+ * A map consisting of full course names and their corresponding abbreviations
+ * from 1996-2023.
+ */
+const fullCourseNames = {
+  "Vertically Integrated Project": "VIP",
+  Wolof: "WOLO",
+  "Electrical & Computer Engr": "ECE",
+  "Computer Science": "CS",
+  "Cooperative Work Assignment": "COOP",
+  "Cross Enrollment": "UCGA",
+  "Earth and Atmospheric Sciences": "EAS",
+  English: "ENGL",
+  "Foreign Studies": "FS",
+  French: "FREN",
+  "Georgia Tech": "GT",
+  "Civil and Environmental Engr": "CEE",
+  "College of Architecture": "COA",
+  "College of Engineering": "COE",
+  "Computational Mod, Sim, & Data": "CX",
+  "Computational Science & Engr": "CSE",
+  "Biological Sciences": "BIOS",
+  Biology: "BIOL",
+  "Biomed Engr/Joint Emory PKU": "BMEJ",
+  "Biomedical Engineering": "BMED",
+  "Industrial Design": "ID",
+  "International Affairs": "INTA",
+  "International Logistics": "IL",
+  Internship: "INTN",
+  "Intl Executive MBA": "IMBA",
+  Japanese: "JAPN",
+  Korean: "KOR",
+  "Learning Support": "LS",
+  Linguistics: "LING",
+  "Literature, Media & Comm": "LMC",
+  Management: "MGT",
+  "Management of Technology": "MOT",
+  "Manufacturing Leadership": "MLDR",
+  "Materials Science & Engr": "MSE",
+  Accounting: "ACCT",
+  "Aerospace Engineering": "AE",
+  "Air Force Aerospace Studies": "AS",
+  "Applied Physiology": "APPH",
+  "Mechanical Engineering": "ME",
+  "Medical Physics": "MP",
+  "Military Science & Leadership": "MSL",
+  "Modern Languages": "ML",
+  Music: "MUSI",
+  "Naval Science": "NS",
+  Neuroscience: "NEUR",
+  Chemistry: "CHEM",
+  Chinese: "CHIN",
+  "City Planning": "CP",
+  Economics: "ECON",
+  "Elect & Comp Engr-Professional": "ECEP",
+  Physics: "PHYS",
+  "Political Science": "POL",
+  "Polymer, Textile and Fiber Eng": "PTFE",
+  Psychology: "PSYC",
+  "Public Policy": "PUBP",
+  "Public Policy/Joint GSU PhD": "PUBJ",
+  Russian: "RUSS",
+  Sociology: "SOC",
+  Spanish: "SPAN",
+  Mathematics: "MATH",
+  "Center Enhancement-Teach/Learn": "CETL",
+  "Chemical & Biomolecular Engr": "CHBE",
+  "Biomedical Engr/Joint Emory": "BMEM",
+  "Bldg Construction-Professional": "BCP",
+  "Building Construction": "BC",
+  Swahili: "SWAH",
+  "Georgia Tech Lorraine": "GTL",
+  German: "GRMN",
+  "Global Media and Cultures": "GMC",
+  "Health Systems": "HS",
+  History: "HIST",
+  "History, Technology & Society": "HTS",
+  "Industrial & Systems Engr": "ISYE",
+  "Nuclear & Radiological Engr": "NRE",
+  Philosophy: "PHIL",
+  "Applied Systems Engineering": "ASE",
+  Arabic: "ARBC",
+  Architecture: "ARCH",
+  "Office of International Educ": "OIE",
+  "College of Sciences": "COS",
+  "Ivan Allen College": "IAC",
+  "Serve, Learn, Sustain": "SLS",
+  Persian: "PERS",
+  Hebrew: "HEBW",
+  Hindi: "HIN",
+  "Int'l Plan Co-op Abroad": "IPCO",
+  "Int'l Plan Intern Abroad": "IPIN",
+  "Int'l Plan-Exchange Program": "IPFS",
+  "Int'l Plan-Study Abroad": "IPSA",
+  Portuguese: "PORT",
+  "Professional Practice": "DOPP",
+  "Lit, Communication & Culture": "LCC",
+  "Health Performance Science": "HPS",
+  "Philosophy of Science/Tech": "PST",
+  "Health Physics": "HP",
+  "Regents' Reading Skills": "RGTR",
+  "Regents' Writing Skills": "RGTE",
+  "Chemical Engineering": "CHE",
+  "Textile & Fiber Engineering": "TFE",
+  "Textile Engineering": "TEX",
+  "Management Science": "MSCI",
+  "Materials Engineering": "MATE",
+  "Civil Engineering": "CE",
+  "Electrical Engineering": "EE",
+  "Computer Engineering": "CMPE",
+  "Military Science": "MS",
+  "Nuclear Engineering": "NE",
+  "Physical Education": "PE",
+  "Engineering Graphics": "EGR",
+  "Engr Science and Mechanics": "ESM",
+  "Technology & Science Policy": "TASP",
+  "Foreign Language": "FL",
+  "Studies Abroad": "SA",
+};
+export const courseMap = new Map(Object.entries(fullCourseNames));
