+++ src/steps/details.ts
@@ -124,6 +124,65 @@ export async function downloadCoursePrereqDetails(
 }
 
 /**
+ * Downloads the corequisites for a single course
+ * @param term - The term string
+ * @param courseId - The joined course id (SUBJECT NUMBER); i.e. `"CS 2340"`
+ */
+export async function downloadCourseCoreqDetails(
+  term: string,
+  courseId: string
+): Promise<string> {
+  const splitResult = splitCourseId(courseId);
+  if (splitResult === null) {
+    warn("could not split course ID; skipping detail scraping", {
+      courseId,
+    });
+    return "";
+  }
+
+  const [subject, number] = splitResult;
+  const parameters = {
+    term,
+    subjectCode: subject,
+    courseNumber: number,
+  };
+  const query = `?${concatParams(parameters)}`;
+  const url = `https://registration.banner.gatech.edu/StudentRegistrationSsb/ssb/courseSearchResults/getCorequisites${query}`;
+
+  // Perform the request in a retry loop
+  // (sometimes, we get rate limits/transport errors so this tries to mitigates them)
+  const maxAttemptCount = 10;
+  try {
+    const response = await backOff(
+      () =>
+        axios.get<string>(url, {
+          headers: {
+            "User-Agent": "gt-scheduler/crawler",
+          },
+        }),
+      {
+        // See https://github.com/coveooss/exponential-backoff for options API
+        jitter: "full",
+        numOfAttempts: maxAttemptCount,
+        retry: (err, attemptNumber) => {
+          error(`an error occurred while fetching details`, err, {
+            courseId,
+            url,
+            attemptNumber,
+            tryingAgain: attemptNumber < maxAttemptCount,
+          });
+          return true;
+        },
+      }
+    );
+    return response.data;
+  } catch (err) {
+    error(`exhausted retries for fetching coreqs`, err, { courseId });
+    throw err;
+  }
+}
+
+/**
  * Downloads restriction information for a single section
  * @param term - The term string
  * @param crn - Course Reference Number (section identifier)
